Struct linera_core::client::ChainClient

pub struct ChainClient<ValidatorNodeProvider, StorageClient> { /* private fields */ }

Client to operate a chain by interacting with validators and the given local storage implementation.

The chain being operated is called the “local chain” or just the “chain”.

As a rule, operations are considered successful (and communication may stop) when they succeeded in gathering a quorum of responses.

Implementations

impl<P, S> ChainClient<P, S>

pub fn new(
chain_id: ChainId,
known_key_pairs: Vec<KeyPair>,
validator_node_provider: P,
storage_client: S,
admin_id: ChainId,
max_pending_messages: usize,
block_hash: Option<CryptoHash>,
timestamp: Timestamp,
next_block_height: BlockHeight,
cross_chain_delay: Duration,
cross_chain_retries: usize
) -> Self

pub fn chain_id(&self) -> ChainId

pub fn block_hash(&self) -> Option<CryptoHash>

Returns the hash of the latest known block.

pub fn timestamp(&self) -> Timestamp

Returns the earliest possible timestamp for the next block.

pub fn next_block_height(&self) -> BlockHeight

pub fn pending_block(&self) -> &Option<Block>

impl<P, S> ChainClient<P, S>where
P: ValidatorNodeProvider + Sync,
S: Store + Clone + Send + Sync + 'static,
ViewError: From<S::ContextError>,

pub async fn chain_state_view(
&self,
chain_id: Option<ChainId>
) -> Result<Arc<ChainStateView<S::Context>>, NodeError>

Obtains a ChainStateView for a given ChainId.

pub async fn epochs(&mut self) -> Result<Vec<Epoch>, NodeError>

Obtains the epochs of the committees trusted by the local chain.

pub async fn local_committee(&mut self) -> Result<Committee, NodeError>

Obtains the committee for the current epoch of the local chain.

pub async fn identity(&mut self) -> Result<Owner, Error>

Obtains the identity of the current owner of the chain. HACK: In the case of a multi-owner chain, we pick one identity for which we know the private key.

pub async fn key_pair(&mut self) -> Result<&KeyPair>

Obtains the key pair associated to the current identity.

pub async fn public_key(&mut self) -> Result<PublicKey>

Obtains the public key associated to the current identity.

pub async fn subscribe_all(
&mut self,
chain_ids: Vec<ChainId>
) -> Result<NotificationStream>

Subscribes to notifications for all validators.

pub async fn transfer(
&mut self,
owner: Option<Owner>,
amount: Amount,
recipient: Recipient,
user_data: UserData
) -> Result<Certificate>

Sends money.

pub async fn claim(
&mut self,
owner: Owner,
target: ChainId,
recipient: Recipient,
amount: Amount,
user_data: UserData
) -> Result<Certificate>

Claims money in a remote chain.

pub async fn execute_operations(
&mut self,
operations: Vec<Operation>
) -> Result<Certificate>

Executes a list of operations.

pub async fn execute_operation(
&mut self,
operation: Operation
) -> Result<Certificate>

Executes an operation.

pub async fn query_application(&mut self, query: &Query) -> Result<Response>

Queries an application.

pub async fn local_balance(&mut self) -> Result<Balance>

pub async fn update_validators_about_local_chain(&mut self) -> Result<()>

Attempts to update all validators about the local chain.

pub async fn transfer_to_account(
&mut self,
owner: Option<Owner>,
amount: Amount,
account: Account,
user_data: UserData
) -> Result<Certificate>

Sends tokens to a chain.

pub async fn burn(
&mut self,
owner: Option<Owner>,
amount: Amount,
user_data: UserData
) -> Result<Certificate>

Burns tokens.

pub async fn synchronize_and_recompute_balance(&mut self) -> Result<Balance>

Attempts to synchronize with validators and re-compute our balance.

pub async fn retry_pending_block(&mut self) -> Result<Option<Certificate>>

Retries the last pending block

pub async fn clear_pending_block(&mut self)

Clears the information on any operation that previously failed.

pub async fn receive_certificate(
&mut self,
certificate: Certificate
) -> Result<()>

Processes confirmed operation for which this chain is a recipient.

pub async fn rotate_key_pair(&mut self, key_pair: KeyPair) -> Result<Certificate>

Rotates the key of the chain.

pub async fn transfer_ownership(
&mut self,
new_public_key: PublicKey
) -> Result<Certificate>

Transfers ownership of the chain.

pub async fn share_ownership(
&mut self,
new_public_key: PublicKey
) -> Result<Certificate>

Adds another owner to the chain.

pub async fn open_chain(
&mut self,
public_key: PublicKey
) -> Result<(EffectId, Certificate)>

Opens a new chain with a derived UID.

pub async fn close_chain(&mut self) -> Result<Certificate>

Closes the chain (and loses everything in it!!).

pub async fn publish_bytecode(
&mut self,
contract: Bytecode,
service: Bytecode
) -> Result<(BytecodeId, Certificate)>

Publishes some bytecode.

pub async fn create_application(
&mut self,
bytecode_id: BytecodeId,
parameters: Vec<u8>,
initialization_argument: Vec<u8>,
required_application_ids: Vec<UserApplicationId>
) -> Result<(UserApplicationId, Certificate)>

Creates an application by instantiating some bytecode.

pub async fn stage_new_committee(
&mut self,
validators: BTreeMap<ValidatorName, ValidatorState>
) -> Result<Certificate>

Creates a new committee and start using it (admin chains only).

pub async fn process_inbox(&mut self) -> Result<Vec<Certificate>>

Creates an empty block to process all incoming messages. This may require several blocks.

pub async fn subscribe_to_new_committees(&mut self) -> Result<Certificate>

Starts listening to the admin chain for new committees. (This is only useful for other genesis chains or for testing.)

pub async fn unsubscribe_from_new_committees(&mut self) -> Result<Certificate>

Stops listening to the admin chain for new committees. (This is only useful for testing.)

pub async fn subscribe_to_published_bytecodes(
&mut self,
chain_id: ChainId
) -> Result<Certificate>

Starts listening to the given chain for published bytecodes.

pub async fn unsubscribe_from_published_bytecodes(
&mut self,
chain_id: ChainId
) -> Result<Certificate>

Stops listening to the given chain for published bytecodes.

pub async fn finalize_committee(&mut self) -> Result<Certificate>

Deprecates all the configurations of voting rights but the last one (admin chains only). Currently, each individual chain is still entitled to wait before accepting this command. However, it is expected that deprecated validators stop functioning shortly after such command is issued.

pub async fn transfer_to_account_unsafe_unconfirmed(
&mut self,
owner: Option<Owner>,
amount: Amount,
account: Account,
user_data: UserData
) -> Result<Certificate>

Sends money to a chain. Do not check balance. (This may block the client) Do not confirm the transaction.