linera_core

Module data_types

Data types exchanged between clients, workers, and validator nodes.

Structs

CertificatesByHeightRequest

Request for downloading certificates by heights.

ChainInfo

Information about a chain, returned in response to a ChainInfoQuery.

ChainInfoQuery

Request information about a chain.

ChainInfoResponse

The response to an ChainInfoQuery

RoundTimeout

The time at which the current round times out, and the round and height it applies to.

ShardInfo

Information about shard allocation for a chain.

Enums

ClientOutcome

The outcome of trying to commit a list of operations to the chain.

CrossChainRequest

An internal request between chains within a validator.