linera_core

Module local_node

Structs

LocalNode

A local node with a single worker, typically used by clients.

LocalNodeClient

A client to a local node.

Enums

LocalNodeError

Error type for the operations on a local node.

Traits

LocalChainInfoExt

Extension trait for ChainInfos from our local node. These should always be valid and contain the requested information.