Trait linera_core::node::LocalValidatorNode

pub trait LocalValidatorNode {
    type NotificationStream: Stream<Item = Notification> + Unpin;

    // Required methods
    async fn handle_block_proposal(
        &mut self,
        proposal: BlockProposal
    ) -> Result<ChainInfoResponse, NodeError>;
    async fn handle_lite_certificate(
        &mut self,
        certificate: LiteCertificate<'_>,
        delivery: CrossChainMessageDelivery
    ) -> Result<ChainInfoResponse, NodeError>;
    async fn handle_certificate(
        &mut self,
        certificate: Certificate,
        hashed_certificate_values: Vec<HashedCertificateValue>,
        hashed_blobs: Vec<HashedBlob>,
        delivery: CrossChainMessageDelivery
    ) -> Result<ChainInfoResponse, NodeError>;
    async fn handle_chain_info_query(
        &mut self,
        query: ChainInfoQuery
    ) -> Result<ChainInfoResponse, NodeError>;
    async fn get_version_info(&mut self) -> Result<VersionInfo, NodeError>;
    async fn subscribe(
        &mut self,
        chains: Vec<ChainId>
    ) -> Result<Self::NotificationStream, NodeError>;
    async fn download_blob(
        &mut self,
        blob_id: BlobId
    ) -> Result<Blob, NodeError>;
    async fn download_certificate_value(
        &mut self,
        hash: CryptoHash
    ) -> Result<HashedCertificateValue, NodeError>;
    async fn download_certificate(
        &mut self,
        hash: CryptoHash
    ) -> Result<Certificate, NodeError>;
    async fn blob_last_used_by(
        &mut self,
        blob_id: BlobId
    ) -> Result<CryptoHash, NodeError>;
}

How to communicate with a validator node.

Required Associated Types

type NotificationStream: Stream<Item = Notification> + Unpin

Required Methods

async fn handle_block_proposal( &mut self, proposal: BlockProposal ) -> Result<ChainInfoResponse, NodeError>

Proposes a new block.

async fn handle_lite_certificate( &mut self, certificate: LiteCertificate<'_>, delivery: CrossChainMessageDelivery ) -> Result<ChainInfoResponse, NodeError>

Processes a certificate without a value.

async fn handle_certificate( &mut self, certificate: Certificate, hashed_certificate_values: Vec<HashedCertificateValue>, hashed_blobs: Vec<HashedBlob>, delivery: CrossChainMessageDelivery ) -> Result<ChainInfoResponse, NodeError>

Processes a certificate.

async fn handle_chain_info_query( &mut self, query: ChainInfoQuery ) -> Result<ChainInfoResponse, NodeError>

Handles information queries for this chain.

async fn get_version_info(&mut self) -> Result<VersionInfo, NodeError>

Gets the version info for this validator node.

async fn subscribe( &mut self, chains: Vec<ChainId> ) -> Result<Self::NotificationStream, NodeError>

Subscribes to receiving notifications for a collection of chains.

async fn download_blob(&mut self, blob_id: BlobId) -> Result<Blob, NodeError>

async fn download_certificate_value( &mut self, hash: CryptoHash ) -> Result<HashedCertificateValue, NodeError>

async fn download_certificate( &mut self, hash: CryptoHash ) -> Result<Certificate, NodeError>

async fn blob_last_used_by( &mut self, blob_id: BlobId ) -> Result<CryptoHash, NodeError>

Returns the hash of the Certificate that last used a blob.

Object Safety

This trait is not object safe.