Trait linera_core::node::ValidatorNode

pub trait ValidatorNode {
    // Required methods
    fn handle_block_proposal<'life0, 'async_trait>(
        &'life0 mut self,
        proposal: BlockProposal
    ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_lite_certificate<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        certificate: LiteCertificate<'life1>
    ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn handle_certificate<'life0, 'async_trait>(
        &'life0 mut self,
        certificate: Certificate,
        blobs: Vec<HashedValue>
    ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_chain_info_query<'life0, 'async_trait>(
        &'life0 mut self,
        query: ChainInfoQuery
    ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 mut self,
        chains: Vec<ChainId>
    ) -> Pin<Box<dyn Future<Output = Result<NotificationStream, NodeError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

How to communicate with a validator or a local node.

Required Methods

fn handle_block_proposal<'life0, 'async_trait>( &'life0 mut self, proposal: BlockProposal ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Proposes a new block.

fn handle_lite_certificate<'life0, 'life1, 'async_trait>( &'life0 mut self, certificate: LiteCertificate<'life1> ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Processes a certificate without a value.

fn handle_certificate<'life0, 'async_trait>( &'life0 mut self, certificate: Certificate, blobs: Vec<HashedValue> ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Processes a certificate.

fn handle_chain_info_query<'life0, 'async_trait>( &'life0 mut self, query: ChainInfoQuery ) -> Pin<Box<dyn Future<Output = Result<ChainInfoResponse, NodeError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Handles information queries for this chain.

fn subscribe<'life0, 'async_trait>( &'life0 mut self, chains: Vec<ChainId> ) -> Pin<Box<dyn Future<Output = Result<NotificationStream, NodeError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Subscribes to receiving notifications for a collection of chains.