linera_core
Module worker

Structs

NetworkActions

Instruct the networking layer to send cross-chain requests and/or push notifications.

Notification

Notification that a chain has a new certified block or a new message.

WorkerState

State of a worker in a validator or a local node.

Enums

Reason

Reason for the notification.

WorkerError

Error type for worker operations.

Traits

ProcessableCertificate