Struct linera_core::worker::WorkerState

pub struct WorkerState<StorageClient> { /* private fields */ }

State of a worker in a validator or a local node.

Implementations

impl<StorageClient> WorkerState<StorageClient>

pub fn new( nickname: String, key_pair: Option<KeyPair>, storage: StorageClient ) -> Self

pub fn new_for_client( nickname: String, storage: StorageClient, recent_values: Arc<Mutex<LruCache<CryptoHash, HashedValue>>>, delivery_notifiers: Arc<Mutex<HashMap<ChainId, BTreeMap<BlockHeight, Vec<Sender<()>>>>>> ) -> Self

pub fn with_allow_inactive_chains(self, value: bool) -> Self

pub fn with_allow_messages_from_deprecated_epochs(self, value: bool) -> Self

pub fn with_grace_period(self, grace_period: Duration) -> Self

Returns an instance with the specified grace period, in microseconds.

Blocks with a timestamp this far in the future will still be accepted, but the validator will wait until that timestamp before voting.

pub fn nickname(&self) -> &str

pub fn storage_client(&self) -> &StorageClient

Returns the storage client so that it can be manipulated or queried by tests in other crates.

impl<StorageClient> WorkerState<StorageClient>where StorageClient: Storage + Clone + Send + Sync + 'static, ViewError: From<StorageClient::ContextError>,

pub async fn fully_handle_certificate( &mut self, certificate: Certificate, blobs: Vec<HashedValue> ) -> Result<ChainInfoResponse, WorkerError>

pub async fn stage_block_execution( &mut self, block: Block ) -> Result<(ExecutedBlock, ChainInfoResponse), WorkerError>

Tries to execute a block proposal without any verification other than block execution.

pub async fn query_application( &mut self, chain_id: ChainId, query: Query ) -> Result<Response, WorkerError>

Executes a Query for an application’s state on a specific chain.

pub async fn cache_recent_value<'a>( &mut self, value: Cow<'a, HashedValue> ) -> bool

pub async fn read_certificate( &self, chain_id: ChainId, height: BlockHeight ) -> Result<Option<Certificate>, WorkerError>

Returns a stored Certificate for a chain’s block.

pub async fn load_application_registry( &self, chain_id: ChainId ) -> Result<ApplicationRegistryView<StorageClient::Context>, WorkerError>

Returns the application registry for a specific chain.

Notes

The returned ApplicationRegistryView holds a lock of the chain it belongs to. Incorrect usage of this method may cause deadlocks.

pub async fn find_incoming_message( &self, chain_id: ChainId, message_id: MessageId ) -> Result<Option<IncomingMessage>, WorkerError>

Returns an IncomingMessage that’s awaiting to be received by the chain specified by chain_id.