Trait linera_core::worker::ValidatorWorker

pub trait ValidatorWorker {
    // Required methods
    fn handle_block_proposal<'life0, 'async_trait>(
        &'life0 mut self,
        proposal: BlockProposal
    ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_lite_certificate<'a, 'life0, 'async_trait>(
        &'life0 mut self,
        certificate: LiteCertificate<'a>,
        notify_message_delivery: Option<Sender<()>>
    ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn handle_certificate<'life0, 'async_trait>(
        &'life0 mut self,
        certificate: Certificate,
        hashed_certificate_values: Vec<HashedCertificateValue>,
        hashed_blobs: Vec<HashedBlob>,
        notify_message_delivery: Option<Sender<()>>
    ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_chain_info_query<'life0, 'async_trait>(
        &'life0 self,
        query: ChainInfoQuery
    ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_cross_chain_request<'life0, 'async_trait>(
        &'life0 mut self,
        request: CrossChainRequest
    ) -> Pin<Box<dyn Future<Output = Result<NetworkActions, WorkerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Interface provided by each physical shard (aka “worker”) of a validator or a local node.

All commands return either the current chain info or an error.

Repeating commands produces no changes and returns no error.

Some handlers may return cross-chain requests, that is, messages to be communicated to other workers of the same validator.

Required Methods

fn handle_block_proposal<'life0, 'async_trait>( &'life0 mut self, proposal: BlockProposal ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Proposes a new block. In case of success, the chain info contains a vote on the new block.

fn handle_lite_certificate<'a, 'life0, 'async_trait>( &'life0 mut self, certificate: LiteCertificate<'a>, notify_message_delivery: Option<Sender<()>> ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

Processes a certificate, e.g. to extend a chain with a confirmed block.

fn handle_certificate<'life0, 'async_trait>( &'life0 mut self, certificate: Certificate, hashed_certificate_values: Vec<HashedCertificateValue>, hashed_blobs: Vec<HashedBlob>, notify_message_delivery: Option<Sender<()>> ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Processes a certificate, e.g. to extend a chain with a confirmed block.

fn handle_chain_info_query<'life0, 'async_trait>( &'life0 self, query: ChainInfoQuery ) -> Pin<Box<dyn Future<Output = Result<(ChainInfoResponse, NetworkActions), WorkerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles information queries on chains.

fn handle_cross_chain_request<'life0, 'async_trait>( &'life0 mut self, request: CrossChainRequest ) -> Pin<Box<dyn Future<Output = Result<NetworkActions, WorkerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles a (trusted!) cross-chain request.