Crate linera_core

This module defines the core Linera protocol.

Re-exports

pub use client::Client;

pub use crate::join_set_ext::JoinSetExt;

pub use crate::join_set_ext::TaskHandle;

pub use environment::wallet;

pub use environment::wallet::Wallet;

pub use environment::Environment;

Modules

client

data_types

environment

join_set_ext

An extension trait to allow determining at compile time how tasks are spawned on the Tokio runtime.

node

notifier

test_utils

worker

Enums

LocalNodeError

Error type for the operations on a local node.

Constants

CHAIN_INFO_MAX_RECEIVED_LOG_ENTRIES

The maximum number of entries in a received_log included in a ChainInfo response.

DEFAULT_QUORUM_GRACE_PERIOD

The default amount of time we wait for additional validators to contribute to the result, as a fraction of how long it took to reach a quorum.