Crate linera_core

This module defines the core Linera protocol.

Re-exports

pub use crate::join_set_ext::JoinSetExt;

pub use crate::join_set_ext::TaskHandle;

Modules

chain_worker

A worker to handle a single chain.

client

data_types

join_set_ext

An extension trait to allow determining at compile time how tasks are spawned on the Tokio runtime.

local_node

node

notifier

remote_node

test_utils

worker