linera_core
Module client

Structs

AbortOnDrop

Wrapper for AbortHandle that aborts when its dropped.

ChainClient

Client to operate a chain by interacting with validators and the given local storage implementation.

ChainClientOptions

Client

A builder that creates ChainClients which share the cache and notifiers.

MessagePolicy

Policies for automatically handling incoming messages.

PendingProposal

A pending proposed block, together with its published blobs.

Unsend

Enums

BlanketMessagePolicy

ChainClientError

Error type for ChainClient.

Functions

create_bytecode_blobs

Creates a compressed Contract, Service and bytecode.

Type Aliases

ChainGuard

ChainGuardMapped

ChainGuardMut