linera_core

Module client

Re-exports

pub use chain_client::ChainClient;

pub use crate::data_types::ClientOutcome;

pub use requests_scheduler::RequestsScheduler;

pub use requests_scheduler::RequestsSchedulerConfig;

pub use requests_scheduler::ScoringWeights;

Modules

chain_client

requests_scheduler

This module manages communication with validator nodes, including connection pooling, load balancing, request deduplication, caching, and performance tracking.

Structs

AbortOnDrop

Wrapper for AbortHandle that aborts when its dropped.

Client

A builder that creates ChainClients which share the cache and notifiers.

PendingProposal

A pending proposed block, together with its published blobs.

Enums

ListeningMode

Defines how we listen to a chain:

TimingType

Statics

DEFAULT_CERTIFICATE_DOWNLOAD_BATCH_SIZE

DEFAULT_SENDER_CERTIFICATE_DOWNLOAD_BATCH_SIZE

Functions

create_bytecode_blobs

Creates a compressed Contract, Service and bytecode.