linera_core

Module data_types

Structs

BlockHeightRange

A range of block heights as used in ChainInfoQuery.

CertificatesByHeightRequest

Request for downloading certificates by heights.

ChainInfo

ChainInfoQuery

Request information about a chain.

ChainInfoResponse

The response to an ChainInfoQuery

RoundTimeout

ShardInfo

Information about shard allocation for a chain.

Enums

ClientOutcome

The outcome of trying to commit a list of operations to the chain.

CrossChainRequest

An internal request between chains within a validator.