Trait linera_core::node::ValidatorNode

pub trait ValidatorNode: Send {
    type NotificationStream: Stream<Item = Notification> + Unpin + Send;

    // Required methods
    fn handle_block_proposal(
        &self,
        proposal: BlockProposal,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn handle_lite_certificate(
        &self,
        certificate: LiteCertificate<'_>,
        delivery: CrossChainMessageDelivery,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn handle_confirmed_certificate(
        &self,
        certificate: GenericCertificate<ConfirmedBlock>,
        blobs: Vec<Blob>,
        delivery: CrossChainMessageDelivery,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn handle_validated_certificate(
        &self,
        certificate: GenericCertificate<ValidatedBlock>,
        blobs: Vec<Blob>,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn handle_timeout_certificate(
        &self,
        certificate: GenericCertificate<Timeout>,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn handle_chain_info_query(
        &self,
        query: ChainInfoQuery,
    ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send;
    fn get_version_info(
        &self,
    ) -> impl Future<Output = Result<VersionInfo, NodeError>> + Send;
    fn get_genesis_config_hash(
        &self,
    ) -> impl Future<Output = Result<CryptoHash, NodeError>> + Send;
    fn subscribe(
        &self,
        chains: Vec<ChainId>,
    ) -> impl Future<Output = Result<Self::NotificationStream, NodeError>> + Send;
    fn download_blob_content(
        &self,
        blob_id: BlobId,
    ) -> impl Future<Output = Result<BlobContent, NodeError>> + Send;
    fn download_certificate(
        &self,
        hash: CryptoHash,
    ) -> impl Future<Output = Result<ConfirmedBlockCertificate, NodeError>> + Send;
    fn download_certificates(
        &self,
        hashes: Vec<CryptoHash>,
    ) -> impl Future<Output = Result<Vec<ConfirmedBlockCertificate>, NodeError>> + Send;
    fn blob_last_used_by(
        &self,
        blob_id: BlobId,
    ) -> impl Future<Output = Result<CryptoHash, NodeError>> + Send;
    fn missing_blob_ids(
        &self,
        blob_ids: Vec<BlobId>,
    ) -> impl Future<Output = Result<Vec<BlobId>, NodeError>> + Send;

}

How to communicate with a validator node.

Required Associated Types

type NotificationStream: Stream<Item = Notification> + Unpin + Send

Required Methods

fn handle_block_proposal( &self, proposal: BlockProposal, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Proposes a new block.

fn handle_lite_certificate( &self, certificate: LiteCertificate<'_>, delivery: CrossChainMessageDelivery, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Processes a certificate without a value.

fn handle_confirmed_certificate( &self, certificate: GenericCertificate<ConfirmedBlock>, blobs: Vec<Blob>, delivery: CrossChainMessageDelivery, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Processes a confirmed certificate.

fn handle_validated_certificate( &self, certificate: GenericCertificate<ValidatedBlock>, blobs: Vec<Blob>, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Processes a validated certificate.

fn handle_timeout_certificate( &self, certificate: GenericCertificate<Timeout>, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Processes a timeout certificate.

fn handle_chain_info_query( &self, query: ChainInfoQuery, ) -> impl Future<Output = Result<ChainInfoResponse, NodeError>> + Send

Handles information queries for this chain.

fn get_version_info( &self, ) -> impl Future<Output = Result<VersionInfo, NodeError>> + Send

Gets the version info for this validator node.

fn get_genesis_config_hash( &self, ) -> impl Future<Output = Result<CryptoHash, NodeError>> + Send

Gets the network’s genesis config hash.

fn subscribe( &self, chains: Vec<ChainId>, ) -> impl Future<Output = Result<Self::NotificationStream, NodeError>> + Send

Subscribes to receiving notifications for a collection of chains.

fn download_blob_content( &self, blob_id: BlobId, ) -> impl Future<Output = Result<BlobContent, NodeError>> + Send

fn download_certificate( &self, hash: CryptoHash, ) -> impl Future<Output = Result<ConfirmedBlockCertificate, NodeError>> + Send

fn download_certificates( &self, hashes: Vec<CryptoHash>, ) -> impl Future<Output = Result<Vec<ConfirmedBlockCertificate>, NodeError>> + Send

Requests a batch of certificates from the validator.

fn blob_last_used_by( &self, blob_id: BlobId, ) -> impl Future<Output = Result<CryptoHash, NodeError>> + Send

Returns the hash of the Certificate that last used a blob.

fn missing_blob_ids( &self, blob_ids: Vec<BlobId>, ) -> impl Future<Output = Result<Vec<BlobId>, NodeError>> + Send

Returns the missing Blobs. by their ids.

Object Safety

This trait is not object safe.