Module linera_core::worker

Structs

NetworkActions

Instruct the networking layer to send cross-chain requests and/or push notifications.

Notification

Notify that a chain has a new certified block or a new message.

WorkerState

State of a worker in a validator or a local node.

Enums

Reason

Reason for the notification.

WorkerError

Error type for ValidatorWorker.

Traits

ValidatorWorker

Interface provided by each physical shard (aka “worker”) of a validator or a local node.